Do not filter the content inside the document header
Jetpack is adding the content inside the document header in metas when the publicize or sharing module are active, we do not need to include the ratings in this part. Moreover to avoid any troubles in case someone is trying to buffer the idea content elsewhere, allow the single idea template parts to be loaded more than once.

Fixes #49